rename rules.ts to effects
